backend/requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
sentence-transformers
faiss-cpu
pydantic
streamlit

